+++ apps/boruta_gateway/lib/boruta_gateway/application.ex
@@ -5,9 +5,7 @@ defmodule BorutaGateway.Application do
 
   use Application
 
-  alias BorutaGateway.Logger
   alias BorutaGateway.Upstreams
-  alias BorutaGateway.Upstreams.ClientSupervisor
 
   @impl Application
   def start(_type, _args) do
@@ -16,11 +14,8 @@ defmodule BorutaGateway.Application do
       %{
         id: Upstreams.Store,
         start: {Upstreams.Store, :start_link, []}
-      },
-      {ClientSupervisor, strategy: :one_for_one}
+      }
     ]
-
-    Logger.start()
 
     children =
       case Application.get_env(:boruta_gateway, :server) do
@@ -33,7 +28,7 @@ defmodule BorutaGateway.Application do
                    [
                      port: Application.fetch_env!(:boruta_gateway, :port),
                      match_function: &Upstreams.match/1,
-                     num_acceptors: 10
+                     num_acceptors: Application.get_env(:boruta_gateway, :num_acceptors, 8)
                    ]
                  ]},
               id: :server
@@ -56,7 +51,7 @@ defmodule BorutaGateway.Application do
                    [
                      port: Application.fetch_env!(:boruta_gateway, :sidecar_port),
                      match_function: &Upstreams.sidecar_match/1,
-                     num_acceptors: 10
+                     num_acceptors: Application.get_env(:boruta_gateway, :num_acceptors, 8)
                    ]
                  ]},
               id: :sidecar_server