+++ apps/boruta_gateway/lib/boruta_gateway/configuration_schemas/gateway.ex
@@ -25,9 +25,16 @@ defmodule BorutaGateway.ConfigurationSchemas.GatewaySchema do
           }
         },
         "strip_uri" => %{"type" => "boolean"},
-        "pool_count" => %{"type" => "number"},
-        "pool_size" => %{"type" => "number"},
-        "max_idle_time" => %{"type" => "number"},
+        "keepalive" => %{"type" => "boolean"},
+        "rate_limit_enabled" => %{"type" => "boolean"},
+        "rate_limit_count" => %{"type" => "number"},
+        "rate_limit_time_unit" => %{
+          "type" => "string",
+          "pattern" => "^(millisecond|second|minute)$"
+        },
+        "rate_limit_penality" => %{"type" => "number"},
+        "rate_limit_timeout" => %{"type" => "number"},
+        "rate_limit_memory_length" => %{"type" => "number"},
         "required_scopes" => %{
           "type" => "object",
           "patternProperties" => %{
@@ -71,9 +78,16 @@ defmodule BorutaGateway.ConfigurationSchemas.GatewaySchema do
           }
         },
         "strip_uri" => %{"type" => "boolean"},
-        "pool_count" => %{"type" => "number"},
-        "pool_size" => %{"type" => "number"},
-        "max_idle_time" => %{"type" => "number"},
+        "keepalive" => %{"type" => "boolean"},
+        "rate_limit_enabled" => %{"type" => "boolean"},
+        "rate_limit_count" => %{"type" => "number"},
+        "rate_limit_time_unit" => %{
+          "type" => "string",
+          "pattern" => "^(millisecond|second|minute)$"
+        },
+        "rate_limit_penality" => %{"type" => "number"},
+        "rate_limit_timeout" => %{"type" => "number"},
+        "rate_limit_memory_length" => %{"type" => "number"},
         "required_scopes" => %{
           "type" => "object",
           "patternProperties" => %{