+++ apps/boruta_gateway/lib/boruta_gateway/upstreams/store.ex
@@ -7,7 +7,6 @@ defmodule BorutaGateway.Upstreams.Store do
 
   alias BorutaGateway.ConfigurationLoader
   alias BorutaGateway.Repo
-  alias BorutaGateway.Upstreams.ClientSupervisor
   alias BorutaGateway.Upstreams.Upstream
 
   def start_link do
@@ -62,42 +61,19 @@ defmodule BorutaGateway.Upstreams.Store do
   end
 
   def handle_call({:match, path_info}, _from, %{upstreams: upstreams} = state) do
-    upstream =
-      with {_prefix_info, upstream} <-
-             Enum.find(upstreams["global"] || [], fn {prefix_info, _upstream} ->
-               path_info = Enum.take(path_info, length(prefix_info))
-
-               Enum.empty?(prefix_info -- path_info)
-             end) do
-        upstream
-      end
-
-    {:reply, upstream, state}
+    {:reply, match_upstream(upstreams["global"] || [], path_info), state}
   end
 
   def handle_call({:sidecar_match, path_info}, _from, %{upstreams: upstreams} = state) do
     node_name = ConfigurationLoader.node_name()
 
-    upstream =
-      with {_prefix_info, upstream} <-
-             Enum.find(upstreams[node_name] || [], fn {prefix_info, _upstream} ->
-               path_info = Enum.take(path_info, length(prefix_info))
-
-               Enum.empty?(prefix_info -- path_info)
-             end) do
-        upstream
-      end
-
-    {:reply, upstream, state}
+    {:reply, match_upstream(upstreams[node_name] || [], path_info), state}
   end
 
   @impl GenServer
   def handle_cast(:hydrate, state) do
     upstreams =
       Repo.all(Upstream)
-      |> Enum.map(fn upstream ->
-        Upstream.with_http_client(upstream)
-      end)
       |> structure()
 
     {:noreply, %{state | hydrated: true, upstreams: upstreams}}
@@ -135,7 +111,6 @@ defmodule BorutaGateway.Upstreams.Store do
         Upstream,
         Enum.map(record, fn {key, value} -> {String.to_atom(key), value} end)
       )
-      |> Upstream.with_http_client()
 
     upstreams
     |> Enum.map(fn {_uri, upstream} -> upstream end)
@@ -155,8 +130,8 @@ defmodule BorutaGateway.Upstreams.Store do
     upstreams
     |> Enum.map(fn {_uri, upstream} -> upstream end)
     |> Enum.map(fn
-      %{id: ^updated_id, http_client: http_client} ->
-        Upstream.with_http_client(%{updated | http_client: http_client})
+      %{id: ^updated_id} ->
+        updated
 
       upstream ->
         upstream
@@ -168,13 +143,8 @@ defmodule BorutaGateway.Upstreams.Store do
     upstreams
     |> Enum.map(fn {_uri, upstream} -> upstream end)
     |> Enum.reject(fn
-      %{id: ^id, http_client: http_client} ->
-        # TODO manage failure
-        true = ClientSupervisor.kill(http_client)
-        true
-
-      _ ->
-        false
+      %{id: ^id} -> true
+      _ -> false
     end)
     |> structure()
   end
@@ -190,5 +160,14 @@ defmodule BorutaGateway.Upstreams.Store do
       |> Enum.sort_by(fn {path_info, _upstream} -> length(path_info) end, :desc)
     end)
     |> Enum.group_by(fn {_path_info, %Upstream{node_name: node_name}} -> node_name end)
+  end
+
+  defp match_upstream(upstreams, path_info) do
+    with {_prefix_info, upstream} <-
+           Enum.find(upstreams, fn {prefix_info, _upstream} ->
+             Enum.take(path_info, length(prefix_info)) == prefix_info
+           end) do
+      upstream
+    end
   end
 end